backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=1.5.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
yfinance>=0.2.0
pytz>=2023.3
